added reducers folder and moved reducer to it. changed reducer name to user-reducer and linked all uses properly
